Extract runtime library
