+++ node/src/parachain/agung_chain_spec.rs
@@ -1,9 +1,13 @@
 use crate::parachain::Extensions;
 use cumulus_primitives_core::ParaId;
 use peaq_agung_runtime::{
-	staking, AccountId, Balance, BalancesConfig, BlockRewardConfig, CouncilConfig, EVMConfig,
+	staking, AccountId, BalancesConfig, BlockRewardConfig, CouncilConfig, EVMConfig,
 	EthereumConfig, GenesisAccount, GenesisConfig, ParachainInfoConfig, ParachainStakingConfig,
-	Precompiles, SudoConfig, SystemConfig, DOLLARS, MILLICENTS, TOKEN_DECIMALS, WASM_BINARY,
+	Precompiles, SudoConfig, SystemConfig, WASM_BINARY,
+};
+use runtime_common::{
+	DOLLARS, MILLICENTS, TOKEN_DECIMALS,
+	Balance,
 };
 use sc_service::{ChainType, Properties};
 use sp_consensus_aura::sr25519::AuthorityId as AuraId;
