+++ runtime/peaq-dev/Cargo.toml
@@ -482,6 +482,11 @@ default-features = false
 path = "../../precompiles/utils"
 default-features = false
 
+[dependencies.runtime-common]
+path = "../common"
+default-features = false
+
+
 [features]
 default = ["std", "aura"]
 aura = []
@@ -547,13 +552,12 @@ std = [
 	'peaq-pallet-rbac/std',
 	'peaq-pallet-rbac-runtime-api/std',
 	'peaq-pallet-storage/std',
+	'peaq-pallet-storage-runtime-api/std',
 	'peaq-pallet-mor/std',
 	'pallet-utility/std',
-	'peaq-pallet-storage-runtime-api/std',
 	'pallet-treasury/std',
 	'pallet-collective/std',
 	'pallet-vesting/std',
- 
 
 	# ETH support
 	'sp-io/std',
@@ -608,6 +612,7 @@ std = [
 	# Customized
 	"peaq-primitives-xcm/std",
 	"precompile-utils/std",
+	"runtime-common/std",
 ]
 
 # Must be enabled for tracing runtimes only
