+++ runtime/peaq/Cargo.toml
@@ -477,6 +477,10 @@ default-features = false
 path = "../../precompiles/utils"
 default-features = false
 
+[dependencies.runtime-common]
+path = "../common"
+default-features = false
+
 
 [features]
 default = ["std", "aura"]
@@ -601,6 +605,7 @@ std = [
 	# Customized
 	"peaq-primitives-xcm/std",
 	"precompile-utils/std",
+	"runtime-common/std",
 ]
 
 # Must be enabled for tracing runtimes only
